requirements.txt
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
Flask==2.0.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysql-connector-python==8.0.26
protobuf==3.17.3
pycparser==2.20
PyJWT==2.1.0
requests==2.26.0
six==1.16.0
urllib3==1.26.6
Werkzeug==2.0.1
gunicorn
